MainScreen Ui Completed Basic App is ready to use

+++ app/src/main/java/com/example/weatherapp/data/models/Forecastday.kt
@@ -7,7 +7,7 @@ import kotlinx.serialization.Serializable
 data class Forecastday(
     val astro: Astro,
     val date: String,
-    val date_epoch: Int,
+    val date_epoch: Long,
     val day: Day,
     val hour: List<Hour>
 )
